# Conflicts:
#	src/config/dev-config.yaml
#	src/service/3DP/VSS.Productivity3D.Service.sln
#	src/service/3DP/src/WebApiModels/VSS.Productivity3D.WebApi.Models.csproj
#	src/service/Filter/AcceptanceTests/tests/ExecutorTests/Internal/TestControllerBase.cs
#	src/service/Filter/AcceptanceTests/tests/ExecutorTests/appsettings.json
#	src/service/Filter/build/yaml/testingvars.env
#	src/service/Filter/src/VSS.Productivity3D.Filter.Common/Executors/DeleteBoundaryExecutor.cs
#	src/service/Filter/src/VSS.Productivity3D.Filter.Common/Executors/DeleteFilterExecutor.cs
#	src/service/Filter/src/VSS.Productivity3D.Filter.Common/Executors/FilterExecutorBase.cs
#	src/service/Filter/src/VSS.Productivity3D.Filter.Common/Executors/GetBoundariesExecutor.cs
#	src/service/Filter/src/VSS.Productivity3D.Filter.Common/Executors/GetBoundaryExecutor.cs
#	src/service/Filter/src/VSS.Productivity3D.Filter.Common/Executors/GetFilterExecutor.cs
#	src/service/Filter/src/VSS.Productivity3D.Filter.Common/Executors/GetFiltersExecutor.cs
#	src/service/Filter/src/VSS.Productivity3D.Filter.Common/Executors/RequestExecutorContainer.cs
#	src/service/Filter/src/VSS.Productivity3D.Filter.Common/Executors/UpsertBoundaryExecutor.cs
#	src/service/Filter/src/VSS.Productivity3D.Filter.Common/Executors/UpsertFilterExecutor.cs
#	src/service/Filter/src/VSS.Productivity3D.Filter.WebApi/Controllers/BoundaryController.cs
#	src/service/Project/AcceptanceTests/VSS.Visionlink.Project.AcceptanceTests.sln
#	src/service/Project/VSS.Visionlink.Project.sln
#	src/service/Project/docker-compose-local.env
#	src/service/Project/src/ProjectWebApiCommon/Executors/RequestExecutorContainer.cs
#	src/service/Project/src/ProjectWebApiCommon/Executors/RequestExecutorContainerFactory.cs
#	src/service/Project/test/UnitTests/MasterDataProjectTests/Executors/ImportFileV4ExecutorTests.cs
#	src/service/TRex/TRex.netstandard.sln

+++ src/service/TRex/src/netstandard/VSS.TRex.Filters/SubGridFilterMasks.cs
@@ -29,8 +29,8 @@ namespace VSS.TRex.Filters
                 return;
             }
 
-            uint originX = SubGridAsLeaf.OriginX;
-            uint originY = SubGridAsLeaf.OriginY;
+            int originX = SubGridAsLeaf.OriginX;
+            int originY = SubGridAsLeaf.OriginY;
 
             double cellSize = SiteModel.CellSize;
 
@@ -77,7 +77,7 @@ namespace VSS.TRex.Filters
 
                 for (byte I = 0; I < SubGridTreeConsts.SubGridTreeDimension; I++)
                 {
-                    uint OriginXPlusI = originX + I;
+                    int OriginXPlusI = originX + I;
                     double CY = OY; // Set to the first row in the column about to be processed
 
                     for (byte J = 0; J < SubGridTreeConsts.SubGridTreeDimension; J++)
@@ -143,7 +143,7 @@ namespace VSS.TRex.Filters
                 // Go over set bits and determine if they are in Design fence boundary
                 FilterMask.ForEachSetBit((X, Y) =>
                 {
-                    SiteModel.Grid.GetCellCenterPosition((uint) (SubGridAsLeaf.OriginX + X), (uint) (SubGridAsLeaf.OriginY + Y), out var CX, out var CY);
+                    SiteModel.Grid.GetCellCenterPosition(SubGridAsLeaf.OriginX + X, SubGridAsLeaf.OriginY + Y, out var CX, out var CY);
                     if (!Filter.SpatialFilter.AlignmentFence.IncludesPoint(CX, CY))
                     {
                         FilterMask.ClearBit(X, Y); // remove interest as its not in design boundary
